+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 ## [vue](./README.md) version `changelog`
 
+##### `v1.2.4`
+- Fixed Chart Timeline
+- Fixed Responsive Table
+- Fixed Chart Vertical size
+
 ##### `v1.2.3`
 - Update Chart Timeline